함수 이동, 코드 정리
