+++ train.py
@@ -12,8 +12,6 @@ from torch.utils.data.dataloader import DataLoader
 from tokenizers import BertWordPieceTokenizer
 
 import modeling
-import modeling2
-import modeling3
 from utils import Config, set_seed, GOOGLE_APPLICATION_CREDENTIAL, MLFLOW_TRACKING_URI
 from data import load_dataset, punctuation, tokenized_dataset
